Done with Posts, Comments, with working auth flow

+++ src/main/java/com/example/blog/service/CommentService.java
@@ -1,13 +1,19 @@
 package com.example.blog.service;
 
-import java.util.Optional;
+import java.util.List;
+import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
+import com.example.blog.dto.CommentRequest;
 import com.example.blog.model.BlogPost;
 import com.example.blog.model.Comment;
 import com.example.blog.model.UserDetail;
@@ -29,8 +35,10 @@ public class CommentService {
     @Autowired
     private UserRepository userRepository;
 
+    private static final Logger logger = LoggerFactory.getLogger(CommentService.class);
+
     @Transactional
-    public Comment createComment (String content, int blogPostId){
+    public CommentRequest createComment (String content, int blogPostId){
         UserDetail user = getAuthenticatedUser();
         BlogPost blogPost = blogPostRepository.findById(blogPostId).orElseThrow(()-> new RuntimeException("Blog not found"));
         System.out.println("abcd"+user);
@@ -38,36 +46,61 @@ public class CommentService {
         comment.setContent(content);
         comment.setUserDetail(user);
         comment.setBlogPost(blogPost);
-        return commentRepository.save(comment);
+        Comment savedComment = commentRepository.save(comment);
+
+        return new CommentRequest(
+            savedComment.getId(),
+            savedComment.getUserDetail().getUserName(),
+            savedComment.getBlogPost(),
+            savedComment.getContent());
     }
 
 
-
-    public java.util.List<Comment> getAllComments(){
-
-        return commentRepository.findAll();
+    public List<CommentRequest> getCommentById(int theId){
+        List<Comment> comments =  commentRepository.findByBlogPostId(theId);
+        return comments.stream().map(comment -> new CommentRequest(comment.getId(),
+         comment.getUserDetail().getUserName(), 
+         comment.getBlogPost(),
+         comment.getContent())).collect(Collectors.toList());
     }
 
-    public Optional<Comment> getCommentById(int theId){
-        return commentRepository.findById(theId);
-    }
-
-    public Comment updateComment(int theId, Comment commentDetails){
+    public ResponseEntity<Comment> updateComment(int theId, Comment commentDetails){
         Comment comment = commentRepository.findById(theId)
         .orElseThrow(()->new RuntimeException("comment not found"));
+
+        UserDetail user = getAuthenticatedUser();
+
+        // check if the authenticate user is the owner of the comment 
+        if(comment.getUserDetail().getId() != user.getId()){
+            String ErrorMessages = "You don't have permisiion to update this comment";
+            logger.error(ErrorMessages);
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+            
+        }
 
         comment.setContent(commentDetails.getContent());
         comment.setUpdatedAt(commentDetails.getUpdatedAt());
 
-        return commentRepository.save(comment);
+        Comment updatedComment = commentRepository.save(comment);
+
+        return ResponseEntity.ok(updatedComment);
 
     }
 
-    public void deleteComment(int theId){
+    public ResponseEntity<Object> deleteComment(int theId){
         Comment comment = commentRepository.findById(theId)
         .orElseThrow(()->new RuntimeException("comment not found"));
+              UserDetail user = getAuthenticatedUser();
 
+        // check if the authenticate user is the owner of the comment 
+        if(comment.getUserDetail().getId() != user.getId()){
+            String ErrorMessages = "You don't have permisiion to update this comment";
+            logger.error(ErrorMessages);
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ErrorMessages);
+            
+        }
         commentRepository.delete(comment);
+        return ResponseEntity.ok("Comment deleted Successfully");
     }
     private UserDetail getAuthenticatedUser() {
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
